Add some extra static files
