+++ packages/lucide-static/.gitignore
@@ -1,3 +1,5 @@
 icons
 lib
 sprite.svg
+tags.json
+icon-nodes.json
